requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
requests>=2.25.0
click>=8.0.0
pandas>=2.3.2
sqlalchemy>=2.0.43
pydantic>=2.11.9
fastapi>=0.117.1
unicorn>=2.1.4
uvicorn>=0.37.0
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
opencv-python>=4.12.0.88
alembic>=1.13.2
